dhcpcd: Store DUID file in a location that can be created
Prior to this change /system/etc/dhcpcd/dhcpcd.duid was being
used, after this change it is now /data/misc/dhcp/dhcpcd.duid.

The code to get a duid first attempts to read this file and if it
does not exist it will then attempt to create it and fail duid
generation if it cannot create the file.

Change-Id: I009cc0b4e4e94fc0adffd0418d0a9b1e3e485f2a
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>